+++ tests/test_indices.py
@@ -3155,16 +3155,16 @@ class TestDrynessIndex:
         np.testing.assert_allclose(di_wet, di_plus_100)
 
 
-class TestDelayedFrostDays:
-    def test_delayed_frost_days(self, tasmin_series):
+class TestLateFrostDays:
+    def test_late_frost_days(self, tasmin_series):
         tasmin = tasmin_series(np.array([-1, 1, 2, -4, 0]) + K2C, start="30/3/2023")
         lat = xr.DataArray(21.45, attrs={"units": "degrees_north"})
         tasmin["lat"] = lat
-        dfd = xci.delayed_frost_days(tasmin)
-        np.testing.assert_allclose(dfd, 1)
+        lfd = xci.late_frost_days(tasmin)
+        np.testing.assert_allclose(lfd, 1)
 
-    def test_delayed_frost_days_lat(self, tasmin_series):
+    def test_late_frost_days_lat(self, tasmin_series):
         tasmin = tasmin_series(np.array([-1, 1, 2, -4, 0]) + K2C, start="30/3/2023")
         lat = xr.DataArray(21.45, attrs={"units": "degrees_north"})
-        dfd = xci.delayed_frost_days(tasmin, lat=lat)
-        np.testing.assert_allclose(dfd, 1)
+        lfd = xci.late_frost_days(tasmin, lat=lat)
+        np.testing.assert_allclose(lfd, 1)