+++ xclim/indicators/atmos/_temperature.py
@@ -79,7 +79,7 @@ __all__ = [
     "huglin_index",
     "biologically_effective_degree_days",
     "latitude_temperature_index",
-    "delayed_frost_days",
+    "late_frost_days",
 ]
 
 
@@ -1221,16 +1221,16 @@ latitude_temperature_index = Temp(
     parameters={"lat": {"kind": InputKind.VARIABLE}, "lat_factor": 60},
 )
 
-delayed_frost_days = TempWithIndexing(
-    title="Delayed frost days",
-    identifier="delayed_frost_days",
+late_frost_days = TempWithIndexing(
+    title="Late frost days",
+    identifier="late_frost_days",
     units="days",
     standard_name="days_with_air_temperature_below_threshold",
-    long_name="Number of delayed days where the daily minimum temperature is below {thresh}",
-    description="{freq} number of delayed days where the daily minimum temperature is below {thresh}."
-    "In the Northern Hemisphere, a frost day is delayed between {start_date_nh} and {end_date_nh}."
-    "In the Southern Hemisphere, a frost day is delayed between {start_date_sh} and {end_date_sh}.",
-    abstract="Number of delayed days where the daily minimum temperature is below a given threshold.",
+    long_name="Number of days where the daily minimum temperature is below {thresh}",
+    description="{freq} number of days where the daily minimum temperature is below {thresh}."
+    "In the Northern Hemisphere, a frost day is late between {start_date_nh} and {end_date_nh}."
+    "In the Southern Hemisphere, a frost day is late between {start_date_sh} and {end_date_sh}.",
+    abstract="Number of days where the daily minimum temperature is below a given threshold.",
     cell_methods="time: sum over days",
-    compute=indices.delayed_frost_days,
+    compute=indices.late_frost_days,
 )